+++ src/client/java/dev/kikugie/elytratrims/ModStatus.java
@@ -1,0 +1,6 @@
+package dev.kikugie.elytratrims;
+
+public class ModStatus {
+    public static boolean stackedTrimsLoaded = false;
+    public static boolean allTheTrimsLoaded = false;
+}